pyproject.toml
[tool]
[tool.poetry]
name = "md2quip"
version = "0.1.0"
homepage = "https://github.com/simonmcc/md2quip"
description = "Mark Down to Quip - publish your docs to Quip."
authors = ["Simon McCartney <simon@mccartney.ie>"]
readme = "README.md"
license =  "Apache-2.0"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "md2quip" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
click = "8.0.1"

black  = { version = "^21.5b2", optional = true}
isort  = { version = "^5.8.0", optional = true}
flake8  = { version = "^3.9.2", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
mypy = {version = "^0.900", optional = true}
pytest  = { version = "^6.2.4", optional = true}
pytest-cov  = { version = "^2.12.0", optional = true}
tox  = { version = "^3.20.1", optional = true}
virtualenv  = { version = "^20.2.2", optional = true}
pip  = { version = "^20.3.1", optional = true}
mkdocs  = { version = "^1.1.2"}
mkdocs-include-markdown-plugin  = { version = "^1.0.0"}
mkdocs-material  = { version = "^6.1.7"}
mkdocstrings  = { version = "^0.15.2"}
mkdocs-material-extensions  = { version = "^1.0.1"}
twine  = { version = "^3.3.0"}
mkdocs-autorefs = {version = "^0.2.1"}
pre-commit = {version = "^2.12.0"}
toml = {version = "^0.10.2"}
bump2version = {version = "^1.0.1"}
PyYAML = "^6.0"
markdown2 = "^2.4.2"
quipclient = "^0.1"
types-PyYAML = "^6.0.5"
click-log = "^0.4.0"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]

[tool.poetry.scripts]
md2quip = 'md2quip.cli:cli'

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
